% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zips.kc}
\alias{zips.kc}
\title{Example ZIP Code Vector.}
\format{A character vector with a length that matches the number of rows of \code{X.kc} (i.e. \code{zips.kc} is of the order of magnitude of
  1000 elements).  Each entry represents the ZIP code of one business.}
\usage{
zips.kc
}
\description{
A vector of ZIP codes.
}
\details{
\code{zips.kc[i]} represents the ZIP code for the restaurant represented in
the \code{i}th row of the \code{X.kc} inspection scores matrix. ZIP codes in
\code{zips.kc} have the format "zip.j" where j is an integer between 1 and 11.
We replaced real King County (WA) ZIP code names with fictional ZIP code
names so as to make data less identifiable (although inspection data is
already publicly available). In this replacement step, we also
demonstrate that our functions can be applied not solely over character
vectors of real ZIP codes, but any vector of character strings representing the
same facet for all restaurants can be used in the grading process.
}
\keyword{datasets}

